+++ nuxeo/__init__.py
@@ -18,7 +18,7 @@ from __future__ import unicode_literals
 
 
 __author__ = 'Nuxeo'
-__version__ = '2.1.1'
+__version__ = '2.1.2'
 __copyright__ = """
     Copyright Nuxeo (https://www.nuxeo.com) and others.