+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 Release date: ``2019-xx-xx``
 
 - `NXPY-102 <https://jira.nuxeo.com/browse/NXPY-102>`__: Set Upload operations to void operations
+- `NXPY-103 <https://jira.nuxeo.com/browse/NXPY-103>`__: Launch flake8 on actual client data
 
 Technical changes
 -----------------